Move config-file store wrappers into each plugin
The centralized withConfigFile in @executor/config needed to understand
every plugin's stored source shape, which forced `as any` casts and a
`config: any` type on the store generic. Replaced with:

- Low-level primitives in @executor/config/helpers: translateSecretHeaders
  (shared by openapi + graphql for secret-ref headers) and
  makeConfigFileBinding (typed Effect wrapper)
- Per-plugin withConfigFile in each plugin's sdk, typed against that
  plugin's concrete store and StoredSource — no casts

Also fix EditOpenApiSource input focus loss and missing secret creation:
extract CustomHeaderRow / InlineCreateSecret / HeaderValuePreview into a
shared custom-header-row.tsx so both Add and Edit use the same form with
inline secret creation. Edit form now tracks headers as an array with
index-based keys instead of a name-keyed record, so editing the name
field no longer unmounts the input.
