+++ packages/plugins/graphql/src/sdk/index.ts
@@ -15,6 +15,7 @@ export {
   makeKvOperationStore,
   makeInMemoryOperationStore,
 } from "./kv-operation-store";
+export { withConfigFile } from "./config-file-store";
 
 export {
   GraphqlIntrospectionError,
