+++ packages/plugins/openapi/package.json
@@ -18,6 +18,7 @@
     "@apidevtools/swagger-parser": "^12.1.0",
     "@effect/platform": "catalog:",
     "@effect/platform-node": "catalog:",
+    "@executor/config": "workspace:*",
     "@executor/sdk": "workspace:*",
     "effect": "catalog:",
     "openapi-types": "^12.1.3",
